VictorMurashev-homework-20240822
